backend-archive/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
js2py
bcrypt
dj-database-url
gunicorn
